+++ CHANGELOG.md
@@ -18,11 +18,15 @@ All notable changes to this project will be documented in this file.
 - Changed alert component to `display: block` by default, no longer requiring class `in`. Close-buttons for the alert component now adds class `d-none` instead.
 - Changed modal component to use class `d-block` instead of class `in`. Open/close-buttons for the modal component now toggles class `d-block` instead.
 - Optimized `px-script` code for rangeslider component. It no longer triggers eventlistener for chrome-related styling in other browsers.
-- Fixed form components for other browsers than chrome.
 
 ### Removed
 
 - Removed class `fade` from modal for now.
+
+### Fixed
+
+- Fixed form components for other browsers than chrome.
+- Fixed bug with breadcrumb component not rendering correctly if using spaces between `<li>`-elements.
 
 ## [0.6.0] - 2018-06-08