+++ CHANGELOG.md
@@ -2,6 +2,33 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.7.0] - 2018-06-?
+
+### Added
+
+- Added well component.
+- Added a styling for using `text-muted` within a text-color modifyer class (i.e. `color-brand`).
+- Added section about breakpoints.
+- Added section about display classes.
+- Added table with supported browsers in browser-support section.
+- Added card component.
+
+### Changed
+
+- Changed the name of the alert class `alert-error` to `alert-danger` to match the other components.
+- Changed alert component to `display: block` by default, no longer requiring class `in`. Close-buttons for the alert component now adds class `d-none` instead.
+- Changed modal component to use class `d-block` instead of class `in`. Open/close-buttons for the modal component now toggles class `d-block` instead.
+- Optimized `px-script` code for rangeslider component. It no longer triggers eventlistener for chrome-related styling in other browsers.
+
+### Removed
+
+- Removed class `fade` from modal for now.
+
+### Fixed
+
+- Fixed form components for other browsers than chrome.
+- Fixed bug with breadcrumb component not rendering correctly if using spaces between `<li>`-elements.
+
 ## [0.6.0] - 2018-06-08
 
 ### Added